Feat: Material Block recognition for Geneva Layout
Includes functionalities to recognize material description blocks for the Geneva borehole profile layout.
The code detects so called layer index columns which indicate the presence of a material description rect as well as their position marks how to split the material description rect into material blocks.
Note: Some visualization code is part of this commit which should be dropped in a follow up commit.